requirements.txt
requests==2.28.1
python-telegram-bot==13.14
python-dotenv==0.21.0
environs==9.5.0
